+++ templates/ui-configs/chakra/components/Input.tsx
@@ -1,0 +1,8 @@
+// @ts-nocheck
+import { Input as ChakraInput, InputProps as ChakraInputProps } from '@chakra-ui/react';
+
+export interface InputProps extends ChakraInputProps {}
+
+export function Input(props: InputProps) {
+  return <ChakraInput {...props} />;
+}