+++ templates/ui-configs/semantic/components/Input.tsx
@@ -1,0 +1,8 @@
+// @ts-nocheck
+import { Input as SemanticInput, InputProps as SemanticInputProps } from 'semantic-ui-react';
+
+export interface InputProps extends SemanticInputProps {}
+
+export function Input(props: InputProps) {
+  return <SemanticInput {...props} />;
+}